+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Improvements
 
 #### Fixes
+- Fix connecting to deluge version < 2.x ([10192](https://github.com/pymedusa/Medusa/pull/10192))
 
 -----